+++ test/integration/verify/controls/aws_sns_topic.rb
@@ -4,6 +4,8 @@ aws_region = attribute(:aws_region, default: '', description: 'The AWS region id
 aws_account_id = attribute(:aws_account_id, default: '', description: 'The AWS account identifier.')
 aws_sns_topic_with_subscription_arn = attribute(:aws_sns_topic_with_subscription_arn, default: '', description: 'The AWS SNS topic ARN.')
 aws_sns_topic_no_subscription_arn = attribute(:aws_sns_topic_no_subscription_arn, default: '', description: 'The SNS topic ARN.')
+aws_sns_topic_with_encryption_arn = attribute(:aws_sns_topic_with_encryption_arn, default: '', description: 'The AWS SNS topic with encryption ARN.')
+kms_master_key_id = attribute(:aws_sns_topic_kms_master_key_id, default: '', description: 'The AWS SNS topic kms master key id.')
 
 control 'aws-sns-topic-1.0' do
 
@@ -18,6 +20,10 @@ control 'aws-sns-topic-1.0' do
   describe aws_sns_topic(arn: aws_sns_topic_no_subscription_arn) do
     it { should exist }
     its('confirmed_subscription_count') { should be_zero }
+  end
+
+  describe aws_sns_topic(arn: aws_sns_topic_with_encryption_arn) do
+    its('kms_master_key_id') { should eq kms_master_key_id }
   end
 
   describe aws_sns_topic(arn: "arn:aws:sns:#{aws_region}:#{aws_account_id}:not-existing-arn-dxeeggchuqdbphgmmqebzzedu") do